requirements.txt
transformers
chromadb
torch 
python-telegram-bot
ollama
python-dotenv
openai
langchain-openai